+++ src/mlpack/base.hpp
@@ -24,6 +24,7 @@
 #include <cmath>
 
 // Next, standard includes.
+#include <any>
 #include <cctype>
 #include <cfloat>
 #include <climits>
@@ -31,7 +32,9 @@
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
+#include <optional>
 #include <stdexcept>
+#include <string_view>
 #include <tuple>
 #include <utility>
 #include <numeric>
@@ -81,23 +84,8 @@
   #endif
 #endif
 
-#if defined(MLPACK_HAVE_CXX17)
-  #include <any>
-  #include <string_view>
-  #define MLPACK_ANY std::any
-  #define MLPACK_ANY_CAST std::any_cast
-  #define MLPACK_STRING_VIEW std::string_view
-#elif defined(_MSC_VER)
-  #error "When using Visual Studio, mlpack should be compiled with /std:c++17 or newer."
-#else
-  // Backport std::any from C+17 to C++11 to replace boost::any.
-  // Use mnmlstc backport implementation only if compiler does not
-  // support C++17.
-  #include <mlpack/core/std_backport/any.hpp>
-  #include <mlpack/core/std_backport/string_view.hpp>
-  #define MLPACK_ANY core::v2::any
-  #define MLPACK_ANY_CAST core::v2::any_cast
-  #define MLPACK_STRING_VIEW core::v2::string_view
+#if !defined(MLPACK_HAVE_CXX17)
+  #error "Need to enable C++17 mode in your compiler"
 #endif
 
 // Now include Armadillo through the special mlpack extensions.