+++ src/mlpack/bindings/markdown/print_doc_functions_impl.hpp
@@ -272,7 +272,7 @@ inline std::string PrintTypeDocs()
   data.required = false;
   data.input = true;
   data.loaded = false;
-  data.value = MLPACK_ANY(int(0));
+  data.value = std::any(int(0));
 
   std::string type = GetPrintableType<int>(data);
   oss << " - `" << type << "`{: #doc_" << ToValidHTMLAnchor(type) << "}: "
@@ -280,7 +280,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(double).name());
   data.cppType = "double";
-  data.value = MLPACK_ANY(double(0.0));
+  data.value = std::any(double(0.0));
 
   type = GetPrintableType<double>(data);
   oss << " - `" << type << "`{: #doc_" << ToValidHTMLAnchor(type) << " }: "
@@ -288,7 +288,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(bool).name());
   data.cppType = "double";
-  data.value = MLPACK_ANY(bool(0.0));
+  data.value = std::any(bool(0.0));
 
   type = GetPrintableType<bool>(data);
   oss << " - `" << type << "`{: #doc_" << ToValidHTMLAnchor(type) << " }: "
@@ -296,7 +296,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(std::string).name());
   data.cppType = "std::string";
-  data.value = MLPACK_ANY(std::string(""));
+  data.value = std::any(std::string(""));
 
   type = GetPrintableType<std::string>(data);
   oss << " - `" << type << "`{: #doc_" << ToValidHTMLAnchor(type) << " }: "
@@ -304,7 +304,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(std::vector<int>).name());
   data.cppType = "std::vector<int>";
-  data.value = MLPACK_ANY(std::vector<int>());
+  data.value = std::any(std::vector<int>());
 
   type = GetPrintableType<std::vector<int>>(data);
   oss << " - `" << type << "`{: #doc_" << ToValidHTMLAnchor(type) << " }: "
@@ -312,7 +312,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(std::vector<std::string>).name());
   data.cppType = "std::vector<std::string>";
-  data.value = MLPACK_ANY(std::vector<std::string>());
+  data.value = std::any(std::vector<std::string>());
 
   type = GetPrintableType<std::vector<std::string>>(data);
   oss << " - `" << type << "`{: " << "#doc_" << ToValidHTMLAnchor(type)
@@ -320,7 +320,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(arma::mat).name());
   data.cppType = "arma::mat";
-  data.value = MLPACK_ANY(arma::mat());
+  data.value = std::any(arma::mat());
 
   type = GetPrintableType<arma::mat>(data);
   oss << " - `" << type << "`{: #doc_" << ToValidHTMLAnchor(type) << " }: "
@@ -328,7 +328,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(arma::Mat<size_t>).name());
   data.cppType = "arma::Mat<size_t>";
-  data.value = MLPACK_ANY(arma::Mat<size_t>());
+  data.value = std::any(arma::Mat<size_t>());
 
   // Make sure that we don't print the same type twice, if the integer matrix
   // type is the same type as the regular matrix type.
@@ -342,7 +342,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(arma::rowvec).name());
   data.cppType = "arma::rowvec";
-  data.value = MLPACK_ANY(arma::rowvec());
+  data.value = std::any(arma::rowvec());
   const std::string& rowType = GetPrintableType<arma::rowvec>(data);
 
   oss << " - `" << rowType << "`{: #doc_" << ToValidHTMLAnchor(rowType)
@@ -350,7 +350,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(arma::Row<size_t>).name());
   data.cppType = "arma::Row<size_t>";
-  data.value = MLPACK_ANY(arma::Row<size_t>());
+  data.value = std::any(arma::Row<size_t>());
   const std::string& urowType = GetPrintableType<arma::Row<size_t>>(data);
 
   if (urowType != rowType)
@@ -361,7 +361,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(arma::vec).name());
   data.cppType = "arma::vec";
-  data.value = MLPACK_ANY(arma::vec());
+  data.value = std::any(arma::vec());
   const std::string& colType = GetPrintableType<arma::vec>(data);
 
   // For some languages there is no distinction between column and row vectors.
@@ -374,7 +374,7 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(arma::Col<size_t>).name());
   data.cppType = "arma::Col<size_t>";
-  data.value = MLPACK_ANY(arma::Col<size_t>());
+  data.value = std::any(arma::Col<size_t>());
   const std::string& ucolType = GetPrintableType<arma::Col<size_t>>(data);
 
   // For some languages there is no distinction between column and row vectors.
@@ -388,7 +388,7 @@ inline std::string PrintTypeDocs()
   data.tname =
       std::string(typeid(std::tuple<data::DatasetInfo, arma::mat>).name());
   data.cppType = "std::tuple<data::DatasetInfo, arma::mat>";
-  data.value = MLPACK_ANY(std::tuple<data::DatasetInfo, arma::mat>());
+  data.value = std::any(std::tuple<data::DatasetInfo, arma::mat>());
 
   type = GetPrintableType<std::tuple<data::DatasetInfo, arma::mat>>(data);
   oss << " - `" << type << "`{: #doc_" << ToValidHTMLAnchor(type) << " }: "
@@ -397,14 +397,14 @@ inline std::string PrintTypeDocs()
 
   data.tname = std::string(typeid(priv::mlpackModel).name());
   data.cppType = "mlpackModel";
-  data.value = MLPACK_ANY(new priv::mlpackModel());
+  data.value = std::any(new priv::mlpackModel());
 
   type = GetPrintableType<priv::mlpackModel*>(data);
   oss << " - `" << type << "`{: #doc_model }: "
       << PrintTypeDoc<priv::mlpackModel*>(data) << std::endl;
 
   // Clean up memory.
-  delete MLPACK_ANY_CAST<priv::mlpackModel*>(data.value);
+  delete std::any_cast<priv::mlpackModel*>(data.value);
 
   oss << "</div>" << std::endl;
   oss << std::endl;