+++ src/mlpack/methods/ann/layer/linear3d_impl.hpp
@@ -85,13 +85,12 @@ Linear3DType<MatType, RegularizerType>::operator=(
 }
 
 template<typename MatType, typename RegularizerType>
-void Linear3DType<MatType, RegularizerType>::SetWeights(
-    typename MatType::elem_type* weightsPtr)
+void Linear3DType<MatType, RegularizerType>::SetWeights(const MatType& weightsIn)
 {
-  MakeAlias(weights, weightsPtr, outSize * this->inputDimensions[0] + outSize,
+  MakeAlias(weights, weightsIn, outSize * this->inputDimensions[0] + outSize,
       1);
-  MakeAlias(weight, weightsPtr, outSize, this->inputDimensions[0]);
-  MakeAlias(bias, weightsPtr + weight.n_elem, outSize, 1);
+  MakeAlias(weight, weightsIn, outSize, this->inputDimensions[0]);
+  MakeAlias(bias, weightsIn, outSize, 1, weight.n_elem);
 }
 
 template<typename MatType, typename RegularizerType>