+++ HISTORY.md
@@ -43,9 +43,17 @@ _????-??-??_
 
   * Fix usage of precompiled headers; remove cotire (#3635).
 
-  * Fix non-working `verbose` option for R bindings (#3691).
+  * Fix non-working `verbose` option for R bindings (#3691), and add global
+    `mlpack.verbose` option (#3706).
 
   * Fix divide-by-zero edge case for LARS (#3701).
+
+  * Templatize `SparseCoding` and `LocalCoordinateCoding` to allow different
+    matrix types (#3709, #3711).
+
+  * Fix handling of unused atoms in `LocalCoordinateCoding` (#3711).
+
+  * Move minimum required C++ version from C++14 to C++17 (#3704).
 
 ## mlpack 4.3.0